requirements.txt
annotated-types==0.7.0
anyio==4.9.0
Flask==3.0.0
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
distro==1.9.0
griffe==1.7.3
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.9.0
mcp==1.8.1
numpy==2.2.5
openai==1.78.1
openai-agents==0.0.14
pip==25.1.1
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydantic-settings==2.9.1
python-dotenv==1.1.0
python-multipart==0.0.20
requests==2.32.3
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.1
sse-starlette==2.3.5
starlette==0.46.2
tqdm==4.67.1
types-requests==2.32.0.20250515
typing_extensions==4.13.2
typing-inspection==0.4.0
urllib3==2.4.0
uvicorn==0.34.2
websockets==15.0.1
firebase-admin==6.4.0